requirements.txt
# Flask y extensiones principales
Flask==3.0.0
Flask-CORS==4.0.0
Flask-RESTful==0.3.10

# Manejo de variables de entorno
python-dotenv==1.0.0

# Validación de datos
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Serialización JSON
marshmallow-jsonapi==0.24.0

# Manejo de errores y logging
Flask-Limiter==3.5.0
structlog==23.2.0

# Seguridad
Flask-Talisman==1.1.0
PyJWT==2.8.0
cryptography==41.0.7

# Testing
#pytest==7.4.3
#pytest-flask==1.3.0
#pytest-cov==4.1.0
#factory-boy==3.3.0

# Desarrollo y debugging
Flask-DebugToolbar==0.13.1
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentación de API
#flasgger==0.9.7.1

# Monitoreo y métricas
#prometheus-flask-exporter==0.23.0

# Cache (opcional)
#Flask-Caching==2.1.0
#redis==5.0.1

# Base de datos local (si necesitas una)
#Flask-SQLAlchemy==3.1.1
#alembic==1.13.0

# Utilidades adicionales
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3

# Para desarrollo
watchdog==3.0.0

#Para Odoo
#pypi-xmlrpc==1.1.0

#Formateo de texto
#unicodedata2==0.0.24